backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Test case application for Brendwall"
authors = ["rodion <rpadakov@yandex.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.1"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.7"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
flake8 = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
